--- src/main/java/com/peeerr/climbing/config/auth/CustomAuthFailureHandler.java
@@ -1,31 +1,0 @@
-package com.peeerr.climbing.config.auth;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.peeerr.climbing.dto.common.ApiResponse;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class CustomAuthFailureHandler extends SimpleUrlAuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
-        ObjectMapper mapper = new ObjectMapper();
-
-        String errorResponse = mapper.writeValueAsString(ApiResponse.of(exception.getMessage()));
-
-        response.setStatus(HttpStatus.UNAUTHORIZED.value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding("UTF-8");
-        response.getWriter().write(errorResponse);
-    }
-
-}
