+++ src/main/java/com/peeerr/climbing/controller/CommentController.java
@@ -1,8 +1,8 @@
 package com.peeerr.climbing.controller;
 
 import com.peeerr.climbing.config.auth.CustomUserDetails;
-import com.peeerr.climbing.dto.comment.request.CommentCreateRequest;
-import com.peeerr.climbing.dto.comment.request.CommentEditRequest;
+import com.peeerr.climbing.dto.comment.CommentCreateRequest;
+import com.peeerr.climbing.dto.comment.CommentEditRequest;
 import com.peeerr.climbing.dto.common.ApiResponse;
 import com.peeerr.climbing.service.CommentService;
 import jakarta.validation.Valid;
