+++ src/main/java/com/peeerr/climbing/domain/like/LikeRepository.java
@@ -1,12 +1,13 @@
 package com.peeerr.climbing.domain.like;
 
+import com.peeerr.climbing.domain.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
 
-    Long countLikeByPostId(Long postId);
+    Long countLikeByPost(Post post);
 
     boolean existsLikeByMemberIdAndPostId(Long memberId, Long postId);
