+++ src/main/java/com/peeerr/climbing/domain/post/CustomPostRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.peeerr.climbing.domain.post;
 
-import com.peeerr.climbing.dto.post.request.PostSearchCondition;
-import com.peeerr.climbing.dto.post.response.PostResponse;
+import com.peeerr.climbing.dto.post.PostSearchCondition;
+import com.peeerr.climbing.dto.post.PostResponse;
 import com.peeerr.climbing.dto.post.response.QPostResponse;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.util.StringUtils;
