+++ src/main/java/com/peeerr/climbing/dto/category/CategoryCreateRequest.java
@@ -1,4 +1,4 @@
-package com.peeerr.climbing.dto.category.request;
+package com.peeerr.climbing.dto.category;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
