+++ src/main/java/com/peeerr/climbing/dto/comment/CommentResponse.java
@@ -1,4 +1,4 @@
-package com.peeerr.climbing.dto.comment.response;
+package com.peeerr.climbing.dto.comment;
 
 import com.peeerr.climbing.domain.comment.Comment;
 import lombok.AccessLevel;
