+++ src/main/java/com/peeerr/climbing/dto/post/PostCreateRequest.java
@@ -1,6 +1,5 @@
-package com.peeerr.climbing.dto.post.request;
+package com.peeerr.climbing.dto.post;
 
-import com.peeerr.climbing.domain.post.Post;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -9,20 +8,20 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class PostEditRequest {
+public class PostCreateRequest {
 
     @Size(min = 1, max = 100)
     @NotBlank(message = "제목을 입력해 주세요.")
     private String title;
 
-    @NotBlank(message = "본문을 입력해 주세요")
+    @NotBlank(message = "본문을 입력해 주세요.")
     private String content;
 
     @NotNull(message = "카테고리를 선택해 주세요.")
     private Long categoryId;
 
-    public static PostEditRequest of(String title, String content, Long categoryId) {
-        return new PostEditRequest(title, content, categoryId);
+    public static PostCreateRequest of(String title, String content, Long categoryId) {
+        return new PostCreateRequest(title, content, categoryId);
     }
 
 }
