+++ src/main/java/com/peeerr/climbing/dto/post/PostDetailResponse.java
@@ -1,9 +1,9 @@
-package com.peeerr.climbing.dto.post.response;
+package com.peeerr.climbing.dto.post;
 
 import com.peeerr.climbing.domain.comment.Comment;
 import com.peeerr.climbing.domain.file.File;
 import com.peeerr.climbing.domain.post.Post;
-import com.peeerr.climbing.dto.comment.response.CommentResponse;
+import com.peeerr.climbing.dto.comment.CommentResponse;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
