+++ src/main/java/com/peeerr/climbing/dto/post/PostResponse.java
@@ -1,4 +1,4 @@
-package com.peeerr.climbing.dto.post.response;
+package com.peeerr.climbing.dto.post;
 
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
