+++ src/main/java/com/peeerr/climbing/service/CommentService.java
@@ -5,8 +5,8 @@ import com.peeerr.climbing.domain.comment.CommentRepository;
 import com.peeerr.climbing.domain.post.Post;
 import com.peeerr.climbing.domain.post.PostRepository;
 import com.peeerr.climbing.domain.user.Member;
-import com.peeerr.climbing.dto.comment.request.CommentCreateRequest;
-import com.peeerr.climbing.dto.comment.request.CommentEditRequest;
+import com.peeerr.climbing.dto.comment.CommentCreateRequest;
+import com.peeerr.climbing.dto.comment.CommentEditRequest;
 import com.peeerr.climbing.exception.constant.ErrorMessage;
 import com.peeerr.climbing.exception.ex.EntityNotFoundException;
 import com.peeerr.climbing.exception.ex.UnauthorizedAccessException;
