+++ src/main/java/com/peeerr/climbing/service/PostService.java
@@ -5,11 +5,11 @@ import com.peeerr.climbing.domain.category.CategoryRepository;
 import com.peeerr.climbing.domain.post.Post;
 import com.peeerr.climbing.domain.post.PostRepository;
 import com.peeerr.climbing.domain.user.Member;
-import com.peeerr.climbing.dto.post.request.PostCreateRequest;
-import com.peeerr.climbing.dto.post.request.PostEditRequest;
-import com.peeerr.climbing.dto.post.request.PostSearchCondition;
-import com.peeerr.climbing.dto.post.response.PostResponse;
-import com.peeerr.climbing.dto.post.response.PostDetailResponse;
+import com.peeerr.climbing.dto.post.PostCreateRequest;
+import com.peeerr.climbing.dto.post.PostEditRequest;
+import com.peeerr.climbing.dto.post.PostSearchCondition;
+import com.peeerr.climbing.dto.post.PostResponse;
+import com.peeerr.climbing.dto.post.PostDetailResponse;
 import com.peeerr.climbing.exception.constant.ErrorMessage;
 import com.peeerr.climbing.exception.ex.EntityNotFoundException;
 import com.peeerr.climbing.exception.ex.UnauthorizedAccessException;
