+++ src/test/java/com/peeerr/climbing/controller/CommentControllerTest.java
@@ -3,8 +3,8 @@ package com.peeerr.climbing.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.peeerr.climbing.config.auth.CustomUserDetails;
 import com.peeerr.climbing.domain.user.Member;
-import com.peeerr.climbing.dto.comment.request.CommentCreateRequest;
-import com.peeerr.climbing.dto.comment.request.CommentEditRequest;
+import com.peeerr.climbing.dto.comment.CommentCreateRequest;
+import com.peeerr.climbing.dto.comment.CommentEditRequest;
 import com.peeerr.climbing.service.CommentService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
