+++ src/test/java/com/peeerr/climbing/controller/MemberControllerTest.java
@@ -3,8 +3,8 @@ package com.peeerr.climbing.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.peeerr.climbing.config.auth.CustomUserDetails;
 import com.peeerr.climbing.domain.user.Member;
-import com.peeerr.climbing.dto.member.request.MemberCreateRequest;
-import com.peeerr.climbing.dto.member.request.MemberEditRequest;
+import com.peeerr.climbing.dto.member.MemberCreateRequest;
+import com.peeerr.climbing.dto.member.MemberEditRequest;
 import com.peeerr.climbing.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
