+++ src/test/java/com/peeerr/climbing/controller/PostControllerTest.java
@@ -3,11 +3,11 @@ package com.peeerr.climbing.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.peeerr.climbing.config.auth.CustomUserDetails;
 import com.peeerr.climbing.domain.user.Member;
-import com.peeerr.climbing.dto.post.request.PostCreateRequest;
-import com.peeerr.climbing.dto.post.request.PostEditRequest;
-import com.peeerr.climbing.dto.post.request.PostSearchCondition;
-import com.peeerr.climbing.dto.post.response.PostDetailResponse;
-import com.peeerr.climbing.dto.post.response.PostResponse;
+import com.peeerr.climbing.dto.post.PostCreateRequest;
+import com.peeerr.climbing.dto.post.PostEditRequest;
+import com.peeerr.climbing.dto.post.PostSearchCondition;
+import com.peeerr.climbing.dto.post.PostDetailResponse;
+import com.peeerr.climbing.dto.post.PostResponse;
 import com.peeerr.climbing.service.PostService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
