+++ src/test/java/com/peeerr/climbing/integration/docs/CommentDocTest.java
@@ -12,8 +12,8 @@ import com.peeerr.climbing.domain.post.Post;
 import com.peeerr.climbing.domain.post.PostRepository;
 import com.peeerr.climbing.domain.user.Member;
 import com.peeerr.climbing.domain.user.MemberRepository;
-import com.peeerr.climbing.dto.comment.request.CommentCreateRequest;
-import com.peeerr.climbing.dto.comment.request.CommentEditRequest;
+import com.peeerr.climbing.dto.comment.CommentCreateRequest;
+import com.peeerr.climbing.dto.comment.CommentEditRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
