+++ src/test/java/com/peeerr/climbing/integration/docs/MemberDocTest.java
@@ -9,8 +9,8 @@ import com.peeerr.climbing.domain.like.LikeRepository;
 import com.peeerr.climbing.domain.post.PostRepository;
 import com.peeerr.climbing.domain.user.Member;
 import com.peeerr.climbing.domain.user.MemberRepository;
-import com.peeerr.climbing.dto.member.request.MemberCreateRequest;
-import com.peeerr.climbing.dto.member.request.MemberEditRequest;
+import com.peeerr.climbing.dto.member.MemberCreateRequest;
+import com.peeerr.climbing.dto.member.MemberEditRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
