+++ src/test/java/com/peeerr/climbing/integration/docs/PostDocTest.java
@@ -11,8 +11,8 @@ import com.peeerr.climbing.domain.post.Post;
 import com.peeerr.climbing.domain.post.PostRepository;
 import com.peeerr.climbing.domain.user.Member;
 import com.peeerr.climbing.domain.user.MemberRepository;
-import com.peeerr.climbing.dto.post.request.PostCreateRequest;
-import com.peeerr.climbing.dto.post.request.PostEditRequest;
+import com.peeerr.climbing.dto.post.PostCreateRequest;
+import com.peeerr.climbing.dto.post.PostEditRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
